reverting id type until template instantiation in tpetra is fixed

+++ packages/Interface/src/Client/DTK_Types.hpp
@@ -47,14 +47,14 @@
 namespace DataTransferKit
 {
 //! Entity id type.
-typedef unsigned long long int EntityId;
+typedef unsigned long int EntityId;
 
 //! Invalid entity id.
 static const EntityId dtk_invalid_entity_id =
     std::numeric_limits<EntityId>::max();
 
 //! Support id type.
-typedef unsigned long long int SupportId;
+typedef unsigned long int SupportId;
 
 //! Invalid support id.
 static const SupportId dtk_invalid_support_id =
